% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausbushfire-package.R
\docType{package}
\name{ausbushfire-package}
\alias{ausbushfire}
\alias{ausbushfire-package}
\title{ausbushfire: Bushfire Explorer: Interactive Analysis of Australian Wildfires}
\description{
Provides an interactive Shiny application and supporting functions to explore, visualize, and analyze Australian bushfire data. The package includes cleaned datasets, reproducible scripts, and a pkgdown-generated documentation website for educational purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2025.github.io/assignment-4-packages-and-shiny-apps-Jessica19940828/}
  \item Report bugs at \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-Jessica19940828/issues}
}

}
\author{
\strong{Maintainer}: HengHsieh Chang \email{jessica830828@gmail.com}

}
\keyword{internal}
